refactor: splits action atom
